change all ints to int64

+++ backend/internal/models/assignment_outline.go
@@ -19,7 +19,7 @@ type AssignmentOutlineWithTemplate struct {
 }
 
 type AssignmentOutline struct {
-	ID              int32      `json:"id,omitempty"`
+	ID              int64      `json:"id,omitempty"`
 	TemplateID      int64      `json:"template_id"`
 	CreatedAt       time.Time  `json:"created_at,omitempty"`
 	ReleasedAt      *time.Time `json:"released_at,omitempty"`
@@ -36,7 +36,7 @@ type AssignmentClassroomID struct {
 
 type AssignmentAcceptRequest struct {
 	OrgName        string `json:"org_name"`
-	OrgID          int    `json:"org_id"`
+	OrgID          int64  `json:"org_id"`
 	SourceRepoName string `json:"repo_name"`
 	AssignmentName string `json:"assignment_name"`
 	AssignmentID   int64  `json:"assignment_id"`
